refractor(rest): change serializers and viewsets
Now `Serializer` and `ViewSet` on `api.chat` are more concise since we
are using `rest_framework` tools.

BREAKING CHANGES: Now `ChatMessageSerializer` does not longer returns a
nested `author` this is relegated to `NestedChatMessageSerializer` in
order to have a correct way to display and treat serializers.

BREAKING CHANGES: Move common validators from
`api.permissions.common.objects` to `api.permissions.common` in order to
be more concise.

Changes to be committed:
	new file:   api/permissions/chat.py
	renamed:    api/permissions/common/objects.py -> api/permissions/common.py
	deleted:    api/permissions/common/__init__.py
	modified:   api/routers/chat.py
	modified:   api/serializers/chat.py
	modified:   api/serializers/registration.py
	new file:   api/validators/__init__.py
	new file:   api/validators/chat.py
	modified:   api/viewsets/chat.py
	modified:   api/viewsets/roleplay.py
	modified:   chat/baker_recipes.py
	modified:   chat/consumers.py
	modified:   poetry.lock
	modified:   tests/api/serializers/test_chat_serializers.py
	modified:   tests/api/viewsets/test_chat_viewsets.py
	modified:   tests/registration/test_views.py