feat(dice): add new check to remove command prefix
Added a new checker to remove `settings.BOT_COMMAND_PREFIX` on message
if this is given. This will allow people to use the tool without
worrying about formatting.

Also added a new `common.enums` to order WebSocket Close Codes.

BREAKING CHANGES: Created a new migration `0008_create_bot_user` that as
its name says will create a user with `DEFAULT_FROM_EMAIL` if this is
not created already. *Not really a breaking change but something to have
in mind.*

Changes to be committed:
	deleted:    bot/migrations/__init__.py
	modified:   common/enums.py
	modified:   oilandrope/dev_settings.py
	new file:   registration/migrations/0008_create_bot_user.py
	modified:   roleplay/utils/dice.py
	modified:   setup.cfg
	modified:   tests/roleplay/utils/test_dice.py