feat(chat): now `ChatConsumer` can handle rolls
Checker if message is a roll or not must be done from `Oil &amp; Rope
Tabletop` since this consumer is just dummy and will perform the given
action.
Also added security in order to check if user is authenticated.

BREAKING CHANGES: Removed requirement `token` from WebSocket. The
authentication will be performed from the website itself.
Also change format and now status is not given, only `type` and
`content` except for specific returns.

TODO: We're gonna need some documentation for this... For real...

Changes to be committed:
	modified:   api/serializers/chat.py
	modified:   api/serializers/common.py
	modified:   chat/consumers.py
	modified:   core/consumers.py
	modified:   tests/chat/test_consumers.py