feature/OAR-30 -  Roleplay Functools (#88)
* feat(dice): add `roll_dice` function

Added `roll_dice` function that executes all the logic in a dice roll
and also added the checker `is_dice_roll` in order to check by regex if
a string is a dice roll.

Changes to be committed:
	modified:   bot/exceptions.py
	new file:   core/exceptions.py
	new file:   roleplay/utils/__init__.py
	new file:   roleplay/utils/dice.py
	new file:   tests/roleplay/utils/__init__.py
	new file:   tests/roleplay/utils/test_dice.py

* ci(drop): drop support for `Python 3.8`

BREAKING CHANGES: We are removing support for `Python 3.8` since typing
is not fully implemented.

Changes to be committed:
	modified:   .github/workflows/django_dev.yml

* fix(test): fix on assertion

I don't kwnow how to sum.

Changes to be committed:
	modified:   tests/roleplay/utils/test_dice.py

* feat(command): add `roll` command

Added `roll` command to bot based on the new `roll_dice` function
feature.

Changes to be committed:
	modified:   bot/bot.py
	modified:   bot/cogs/__init__.py
	modified:   bot/cogs/misc.py

* tests(bot): add tests for `roll` command

Added tests for `rol` command using `dpytest`.

Changes to be committed:
	modified:   bot/cogs/__init__.py
	modified:   bot/cogs/misc.py
	new file:   bot/cogs/rol.py
	modified:   tests/bot/cogs/test_misc.py
	modified:   tests/chat/test_models.py

* feat(api): add `roll_dice` function to API

Since now `roll_dice` function is available we added to the API so it
can be called by `XHR` by **authenticated users**.

Changes to be committed:
	modified:   api/urls.py
	modified:   api/views.py
	modified:   tests/api/test_cors.py
	modified:   tests/api/test_views.py

* refactor(imports): using `get_user_model`

Instead of `contants.USER_MODEL` we now use `get_user_model` from
`django.contrib.auth`.

Changes to be committed:
	modified:   api/serializers/registration.py
	modified:   api/viewsets/registration.py
	modified:   api/viewsets/roleplay.py
	modified:   oilandrope/dev_settings.py
	modified:   registration/baker_recipes.py
	modified:   registration/forms/forms.py
	modified:   tests/api/serializers/test_registration_serializers.py
	modified:   tests/api/serializers/test_roleplay_serializer.py
	modified:   tests/api/test_cors.py
	modified:   tests/api/viewsets/test_chat_viewsets.py
	modified:   tests/api/viewsets/test_dynamic_menu_viewsets.py
	modified:   tests/api/viewsets/test_mixins.py
	modified:   tests/api/viewsets/test_registration_viewsets.py
	modified:   tests/api/viewsets/test_roleplay_viewsets.py

* feat(dice): add new check to remove command prefix

Added a new checker to remove `settings.BOT_COMMAND_PREFIX` on message
if this is given. This will allow people to use the tool without
worrying about formatting.

Also added a new `common.enums` to order WebSocket Close Codes.

BREAKING CHANGES: Created a new migration `0008_create_bot_user` that as
its name says will create a user with `DEFAULT_FROM_EMAIL` if this is
not created already. *Not really a breaking change but something to have
in mind.*

Changes to be committed:
	deleted:    bot/migrations/__init__.py
	modified:   common/enums.py
	modified:   oilandrope/dev_settings.py
	new file:   registration/migrations/0008_create_bot_user.py
	modified:   roleplay/utils/dice.py
	modified:   setup.cfg
	modified:   tests/roleplay/utils/test_dice.py

* feat(chat): now `ChatConsumer` can handle rolls

Checker if message is a roll or not must be done from `Oil &amp; Rope
Tabletop` since this consumer is just dummy and will perform the given
action.
Also added security in order to check if user is authenticated.

BREAKING CHANGES: Removed requirement `token` from WebSocket. The
authentication will be performed from the website itself.
Also change format and now status is not given, only `type` and
`content` except for specific returns.

TODO: We're gonna need some documentation for this... For real...

Changes to be committed:
	modified:   api/serializers/chat.py
	modified:   api/serializers/common.py
	modified:   chat/consumers.py
	modified:   core/consumers.py
	modified:   tests/chat/test_consumers.py

* fix(tests): new migration adjustment

Since migration `0008_create_bot_user` we're not gonna get an empty list
of data for this serializer.

Either way it doesn't really make sense to test things that are already
tested by the DRF team.

Also fixed problem with new features on `common.consumers`.

Changes to be committed:
	modified:   core/consumers.py
	modified:   tests/api/serializers/test_registration_serializers.py
	modified:   tests/core/test_consumers.py

* test(cov): increase coverage

Increased coverage by testing some new features on `core.consumers`.

Changes to be committed:
	modified:   core/consumers.py
	modified:   tests/core/test_consumers.py

* test(consumer): consumer testing

Add tests for `TypedConsumerMixin` and `HandlerJsonWebsocketConsumer`.

Changes to be committed:
	modified:   tests/core/test_consumers.py